Refactoring for Http sources
